edge/modules/smart-factory-ml/requirements.txt
azure-iot-device==2.12.0
onnxruntime==1.17.0
numpy==1.24.3
scipy==1.11.1
pandas==2.0.3
flask==2.3.2
aiohttp==3.8.5
pydantic==2.4.2
loguru==0.7.2
python-dotenv==1.0.0